refactor to seperate the views into their own folder and update namespace references
