+++ CookingBook/Views/AboutPage.xaml.cs
@@ -1,4 +1,4 @@
-namespace CookingBook;
+namespace CookingBook.Views;
 
 public partial class AboutPage : ContentPage
 {
