+++ CookingBook/Views/NotePage.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="CookingBook.NotePage"
+             x:Class="CookingBook.Views.NotePage"
              Title="Note">
     <VerticalStackLayout Spacing="10" Margin="5">
         <Editor x:Name="TextEditor"
